[PostInstall] Support the project_directory option

+++ lib/cocoapods-rome/post_install.rb
@@ -1,9 +1,9 @@
 Pod::HooksManager.register('cocoapods-rome', :post_install) do |installer_context|
-  sandbox_root = installer_context.sandbox_root
+  sandbox_root = Pathname(installer_context.sandbox_root)
   sandbox = Pod::Sandbox.new(sandbox_root)
 
-  build_dir = Pathname('build')
-  destination = Pathname('Rome')
+  build_dir = sandbox_root.parent + 'build'
+  destination = sandbox_root.parent + 'Rome'
 
   Pod::UI.puts 'Building frameworks'
 
@@ -15,16 +15,18 @@ Pod::HooksManager.register('cocoapods-rome', :post_install) do |installer_contex
     end
   end
 
-  frameworks = Pathname.glob('build/Release*/Pods*/*.framework')
-  frameworks_by_target = frameworks.group_by { |f| f.to_s =~ %r{^build/Release[^/]*/Pods-?([^/]*)/} && destination + $1 }
+  raise Pod::Informative 'The build directory was not found in the expected location.' unless build_dir.directory?
+
+  frameworks = Pathname.glob(build_dir + 'Release*/Pods*/*.framework')
+  frameworks_by_target = frameworks.group_by { |f| f.to_s =~ %r{build/Release[^/]*/Pods-?([^/]*)/} && destination + $1 }
 
   Pod::UI.puts "Built #{frameworks.count} #{'frameworks'.pluralize(frameworks.count)}"
-  Pod::UI.puts "Copying frameworks to `#{destination}`"
+  Pod::UI.puts "Copying frameworks to `#{destination.relative_path_from Pathname.pwd}`"
 
   destination.rmtree if destination.directory?
-  frameworks_by_target.each do |dest, framework|
+  frameworks_by_target.each do |dest, frameworks|
     FileUtils.mkdir_p dest
-    FileUtils.mv framework, dest
+    FileUtils.mv frameworks, dest
   end
   build_dir.rmtree
 end
